# anyflix-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media Streaming Backend Service",
...
    "contact": {
      "name": "Media Streaming Backend API",
      "url": "https://github.com/your-repo/media-backend"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/sources/": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "📋 List Available Sources",
        "description": "Get all available media sources.",
        "operationId": "get_sources_sources__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/sources/{source}/preferences": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "📋 Get Source Configuration",
        "description": "Get configuration preferences for a specific source.",
        "operationId": "get_source_preferences_sources__source__preferences_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/{source}/popular": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "🔍 Get Popular Content",
        "description": "Get popular content with optional metadata enrichment.",
        "operationId": "get_popular_sources__source__popular_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_SearchResult_"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/{source}/latest": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "🔍 Get Latest Updates",
        "description": "Get latest updates with optional metadata enrichment.",
        "operationId": "get_latest_updates_sources__source__latest_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_SearchResult_"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/{source}/search": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "🔍 Search Content",
        "description": "Search for content with optional metadata enrichment.",
        "operationId": "search_content_sources__source__search_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_SearchResult_"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/{source}/videos": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "🎬 Get Video Streaming Links",
        "description": "Get video sources.",
        "operationId": "get_video_sources_sources__source__videos_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/trailer": {
      "post": {
        "tags": [
          "media-api"
        ],
        "summary": "🎬 Extract Streamable Trailer URL",
        "description": "Extract streamable URL from AniList or TMDB trailer data.\n\nTakes trailer information from AniList or TMDB responses, builds the full YouTube URL,\nand uses ytdlp_extractor to get the actual streamable URL.\n\nArgs:\n    request: TrailerRequest containing either anilist_trailer or tmdb_trailer data\n\nReturns:\n    TrailerResponse with streamable URL and metadata",
        "operationId": "extract_trailer_url_sources_trailer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrailerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MatchSource": {
        "type": "string",
        "enum": [
          "tmdb",
          "anilist"
        ],
        "title": "MatchSource",
        "description": "Match source enumeration."
      },
...
      "MediaInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cover_image_url": {
            "type": "string",
            "title": "Cover Image Url"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/Episode"
            },
            "type": "array",
            "title": "Episodes"
          },
          "seasons_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seasons Length",
            "description": "Number of episodes"
          },
          "alternative_titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternative Titles",
            "description": "Alternative titles in different languages"
          },
          "start_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Year",
            "description": "Start year of the series"
          },
          "end_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Year",
            "description": "End year of the series"
          },
          "fsk_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fsk Rating",
            "description": "FSK age rating (German rating system)"
          },
          "imdb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imdb Id",
            "description": "IMDB ID (e.g., 'tt36469298')"
          },
          "country_of_origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Of Origin",
            "description": "Country of origin"
          },
          "main_genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Genre",
            "description": "Primary genre classification"
          },
          "directors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Directors",
            "description": "List of directors"
          },
          "actors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actors",
            "description": "List of main actors"
          },
          "producers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Producers",
            "description": "List of production companies"
          },
          "backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Url",
            "description": "Backdrop/banner image URL"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id",
            "description": "Internal series ID from the provider"
          }
        },
        "type": "object",
        "required": [
          "name",
          "cover_image_url",
          "description"
        ],
        "title": "MediaInfo",
        "description": "Detailed Media information."
      },
...
      "PaginatedResponse_SearchResult_": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "list": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "List"
          },
          "has_next_page": {
            "type": "boolean",
            "title": "Has Next Page",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "type",
          "list"
        ],
        "title": "PaginatedResponse[SearchResult]"
      },
      "PreferencesResponse": {
        "properties": {
          "preferences": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preferences",
            "description": "Source configuration preferences"
          }
        },
        "type": "object",
        "required": [
          "preferences"
        ],
        "title": "PreferencesResponse",
        "description": "Response for source preferences/configuration."
      },
...
      "SearchResult": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "media_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "best_match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TMDBSearchResult"
              },
              {
                "$ref": "#/components/schemas/Media"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Match"
          },
          "best_match_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "name",
          "image_url",
          "link"
        ],
        "title": "SearchResult",
        "description": "Search result item."
      },
...
      "SourcesResponse": {
        "properties": {
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "List of available source names"
          }
        },
        "type": "object",
        "required": [
          "sources"
        ],
        "title": "SourcesResponse",
        "description": "Response for available sources list."
      },
...
      "TrailerRequest": {
        "properties": {
          "anilist_trailer": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anilist Trailer"
          },
          "tmdb_trailer": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Trailer"
          }
        },
        "type": "object",
        "title": "TrailerRequest",
        "description": "Request model for trailer extraction."
      },
      "TrailerResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "original_url": {
            "type": "string",
            "title": "Original Url"
          },
          "streamable_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Streamable Url"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success",
          "original_url"
        ],
        "title": "TrailerResponse",
        "description": "Response for trailer extraction."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoListResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/VideoSource"
            },
            "type": "array",
            "title": "Videos"
          }
        },
        "type": "object",
        "required": [
          "type",
          "videos"
        ],
        "title": "VideoListResponse",
        "description": "Response for video sources."
      },
      "VideoSource": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "original_url": {
            "type": "string",
            "title": "Original Url"
          },
          "quality": {
            "type": "string",
            "title": "Quality"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "requires_proxy": {
            "type": "boolean",
            "title": "Requires Proxy",
            "default": false
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "subtitles": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitles"
          },
          "audios": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audios"
          }
        },
        "type": "object",
        "required": [
          "url",
          "original_url",
          "quality"
        ],
        "title": "VideoSource",
        "description": "Video source information."
      },
...
    }
  },
...
}